HSEARCH-3269 Move and adapt batch JSR-352 integration tests
