+++ integrationtest/batch-jsr352/pom.xml
@@ -5,7 +5,6 @@
         <groupId>org.hibernate.search</groupId>
         <artifactId>hibernate-search-integrationtest</artifactId>
         <version>6.0.0-SNAPSHOT</version>
-        <relativePath>..</relativePath>
     </parent>
     <artifactId>hibernate-search-integrationtest-batch-jsr352</artifactId>
 
@@ -15,7 +14,13 @@
     <dependencies>
         <dependency>
             <groupId>org.hibernate.search</groupId>
-            <artifactId>hibernate-search-mapper-orm</artifactId>
+            <artifactId>hibernate-search-batch-jsr352-core</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.hibernate.search</groupId>
+            <artifactId>hibernate-search-batch-jsr352-core</artifactId>
+            <type>test-jar</type>
             <scope>test</scope>
         </dependency>
         <dependency>
@@ -23,6 +28,18 @@
             <artifactId>hibernate-search-util-internal-integrationtest-mapper-orm</artifactId>
             <scope>test</scope>
         </dependency>
+
+        <dependency>
+            <groupId>org.hibernate.search</groupId>
+            <artifactId>hibernate-search-backend-lucene</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.hibernate.search</groupId>
+            <artifactId>hibernate-search-backend-elasticsearch</artifactId>
+            <scope>test</scope>
+        </dependency>
+
         <!--
             Leave this dependency here, not in a utils module, so that we don't need to recompile
             the utils to re-run the tests with a different database.
@@ -32,9 +49,49 @@
             <artifactId>${jdbc.driver.artifactId}</artifactId>
             <scope>test</scope>
         </dependency>
+
+        <dependency>
+            <groupId>com.ibm.jbatch</groupId>
+            <artifactId>com.ibm.jbatch-runtime</artifactId>
+            <version>${version.com.ibm.jbatch}</version>
+            <scope>test</scope>
+        </dependency>
+        <!--
+            JBatch doesn't provide a transitive dependency to the JSR-352 API, go figure.
+         -->
+        <dependency>
+            <groupId>javax.batch</groupId>
+            <artifactId>javax.batch-api</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <!--
+            JBatch doesn't provide a transitive dependency to javax.inject, go figure.
+         -->
+        <dependency>
+            <groupId>javax.inject</groupId>
+            <artifactId>javax.inject</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <!--
+            JBatch requires a database in order to work, and it seems it uses SQL that won't work with H2.
+            Anyway, it uses an embedded Derby instance by default, so we just put the Derby driver in the classpath
+            so it won't complain.
+         -->
+        <dependency>
+            <groupId>org.apache.derby</groupId>
+            <artifactId>derby</artifactId>
+            <version>${version.org.apache.derby}</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
     <build>
+        <testResources>
+            <testResource>
+                <filtering>true</filtering>
+                <directory>src/test/resources</directory>
+            </testResource>
+        </testResources>
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
