+++ integrationtest/batch-jsr352/src/test/resources/META-INF/batch.xml
