+++ integrationtest/batch-jsr352/src/test/resources/META-INF/persistence.xml
@@ -1,0 +1,90 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+ ~ Hibernate Search, full-text search for your domain model
+ ~
+ ~ License: GNU Lesser General Public License (LGPL), version 2.1 or later
+ ~ See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
+  -->
+<persistence xmlns="http://xmlns.jcp.org/xml/ns/persistence" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    version="2.1" xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/persistence http://xmlns.jcp.org/xml/ns/persistence/persistence_2_1.xsd">
+    <persistence-unit name="lucene_pu" transaction-type="RESOURCE_LOCAL">
+        <provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
+        <properties>
+            <property name="hibernate.dialect" value="${db.dialect}" />
+            <property name="hibernate.connection.driver_class" value="${jdbc.driver}" />
+            <property name="hibernate.connection.url" value="${jdbc.url}" />
+            <property name="hibernate.connection.user" value="${jdbc.user}" />
+            <property name="hibernate.connection.password" value="${jdbc.pass}" />
+            <property name="hibernate.hbm2ddl.auto" value="create-drop" />
+            <property name="hibernate.session_factory_name" value="primary_session_factory" />
+            <property name="hibernate.session_factory_name_is_jndi" value="false" />
+
+            <property name="hibernate.search.automatic_indexing.strategy" value="none"/>
+            <property name="hibernate.search.backend.type" value="lucene" />
+            <property name="hibernate.search.backend.directory.type" value="local-heap" />
+        </properties>
+    </persistence-unit>
+    <persistence-unit name="elasticsearch_pu" transaction-type="RESOURCE_LOCAL">
+        <provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
+        <properties>
+            <property name="hibernate.dialect" value="${db.dialect}" />
+            <property name="hibernate.connection.driver_class" value="${jdbc.driver}" />
+            <property name="hibernate.connection.url" value="${jdbc.url}" />
+            <property name="hibernate.connection.user" value="${jdbc.user}" />
+            <property name="hibernate.connection.password" value="${jdbc.pass}" />
+            <property name="hibernate.hbm2ddl.auto" value="create-drop" />
+            <property name="hibernate.session_factory_name" value="primary_session_factory" />
+            <property name="hibernate.session_factory_name_is_jndi" value="false" />
+
+            <property name="hibernate.search.automatic_indexing.strategy" value="none"/>
+            <property name="hibernate.search.schema_management.strategy" value="drop-and-create-and-drop"/>
+
+            <property name="hibernate.search.backend.hosts" value="${test.elasticsearch.connection.hosts}" />
+            <property name="hibernate.search.backend.username" value="${test.elasticsearch.connection.username}" />
+            <property name="hibernate.search.backend.password" value="${test.elasticsearch.connection.password}" />
+            <property name="hibernate.search.backend.aws.signing.enabled" value="${test.elasticsearch.connection.aws.signing.enabled}" />
+            <property name="hibernate.search.backend.aws.region" value="${test.elasticsearch.connection.aws.region}" />
+            <property name="hibernate.search.backend.aws.credentials.type" value="${test.elasticsearch.connection.aws.credentials.type}" />
+            <property name="hibernate.search.backend.aws.credentials.access_key_id" value="${test.elasticsearch.connection.aws.credentials.access_key_id}" />
+            <property name="hibernate.search.backend.aws.credentials.secret_access_key" value="${test.elasticsearch.connection.aws.credentials.secret_access_key}" />
+
+            <property name="hibernate.search.backend.schema_management.minimal_required_status" value="yellow" />
+            <property name="hibernate.search.backend.log.json_pretty_printing" value="true" />
+        </properties>
+    </persistence-unit>
+    <persistence-unit name="unused_pu" transaction-type="RESOURCE_LOCAL">
+        <provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
+        <properties>
+            <property name="hibernate.dialect" value="${db.dialect}" />
+            <property name="hibernate.connection.driver_class" value="${jdbc.driver}" />
+            <property name="hibernate.connection.url" value="${jdbc.url}" />
+            <property name="hibernate.connection.user" value="${jdbc.user}" />
+            <property name="hibernate.connection.password" value="${jdbc.pass}" />
+            <property name="hibernate.hbm2ddl.auto" value="create-drop" />
+            <property name="hibernate.session_factory_name" value="unused_session_factory" />
+            <property name="hibernate.session_factory_name_is_jndi" value="false" />
+
+            <property name="hibernate.search.backend.type" value="lucene" />
+            <property name="hibernate.search.backend.directory.type" value="local-heap" />
+        </properties>
+    </persistence-unit>
+    <persistence-unit name="lucene_multiTenancy_pu" transaction-type="RESOURCE_LOCAL">
+        <provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
+        <properties>
+            <property name="hibernate.dialect" value="${db.dialect}" />
+            <property name="hibernate.connection.driver_class" value="${jdbc.driver}" />
+            <property name="hibernate.connection.url" value="${jdbc.url}" />
+            <property name="hibernate.connection.user" value="${jdbc.user}" />
+            <property name="hibernate.connection.password" value="${jdbc.pass}" />
+            <property name="hibernate.hbm2ddl.auto" value="create-drop" />
+            <property name="hibernate.session_factory_name" value="primary_session_factory" />
+            <property name="hibernate.session_factory_name_is_jndi" value="false" />
+            <property name="hibernate.multiTenancy" value="DATABASE" />
+
+            <property name="hibernate.search.automatic_indexing.strategy" value="none" />
+            <property name="hibernate.search.backend.type" value="lucene" />
+            <property name="hibernate.search.backend.directory.type" value="local-heap" />
+            <property name="hibernate.search.backend.multi_tenancy.strategy" value="discriminator" />
+        </properties>
+    </persistence-unit>
+</persistence>
