+++ integrationtest/batch-jsr352/src/test/resources/META-INF/services/batch-services.properties
