--- legacy/integrationtest/jsr352/src/test/java/org/hibernate/search/jsr352/massindexing/test/entity/CompanyGroup.java
@@ -1,58 +1,0 @@
-/*
- * Hibernate Search, full-text search for your domain model
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later
- * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
- */
-package org.hibernate.search.jsr352.massindexing.test.entity;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.ManyToMany;
-
-import org.hibernate.search.annotations.DocumentId;
-import org.hibernate.search.annotations.Indexed;
-import org.hibernate.search.annotations.IndexedEmbedded;
-
-/**
- * @author Yoann Rodiere
- */
-@Entity
-@Indexed
-public class CompanyGroup {
-
-	@Id
-	@DocumentId
-	private String id;
-
-	@ManyToMany
-	@IndexedEmbedded
-	private List<Company> companies = new ArrayList<>();
-
-	public CompanyGroup() {
-	}
-
-	public CompanyGroup(String id, Company ... companies) {
-		this.id = id;
-		Collections.addAll( this.companies, companies );
-	}
-
-	public String getId() {
-		return id;
-	}
-
-	public void setId(String id) {
-		this.id = id;
-	}
-
-	public List<Company> getCompanies() {
-		return companies;
-	}
-
-	public void setCompanies(List<Company> companies) {
-		this.companies = companies;
-	}
-}
