--- legacy/integrationtest/jsr352/src/test/resources/hibernate.properties
@@ -1,18 +1,0 @@
-################################################################################################
-# Hibernate Search, full-text search for your domain model                                     #
-#                                                                                              #
-# License: GNU Lesser General Public License (LGPL), version 2.1 or later                      #
-# See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.  #
-################################################################################################
-
-# The content of this file is only used in
-# org.hibernate.search.jsr352.massindexing.MassIndexingJobWithMultiTenancyIT
-# (to build a SessionFactory, but also to make the @RequiresDialect annotation work).
-# Other tests use the configuration defined in META-INF/persistence.xml.
-
-hibernate.dialect ${db.dialect}
-hibernate.connection.driver_class ${jdbc.driver}
-hibernate.connection.url ${jdbc.url}
-hibernate.connection.username ${jdbc.user}
-hibernate.connection.password ${jdbc.pass}
-hibernate.connection.isolation ${jdbc.isolation}
