+++ legacy/pom.xml
@@ -25,10 +25,8 @@
         <module>legacy</module>
         <module>engine</module>
         <module>orm</module>
-        <module>jsr352</module>
         <module>testing</module>
         <module>infinispan</module>
-        <module>integrationtest/jsr352</module>
         <module>documentation</module>
     </modules>
